requirements.txt
pyarrow
torch~=2.0.1
torchvision~=0.15.0
pyspark==3.3.0
ruff
pyclean
jupyter
numpy~=1.26.4
pandas
pillow~=11.1.0
requests~=2.32.3
diagrams~=0.24.1
scikit-image~=0.24.0
confluent-kafka~=2.8.0